Add EventListener for AuthenticationSuccessEvents
Now every time a user logs in successfully it is checked whether the user
exists or not. And if the user does not exist a new one is created and saved
in the database.
With this the getLoggedInUser() function could be simplified which closes #17
as it now only returns the current user and no longer create one in case a
UserNotFoundException was thrown.