Fix getLoggedInUser() potentially throwing a ClassCastException
In the process of fixing the method it has been rename to `visitingUserId()` and
now returns the id of the user. If a user is not logged in the id is
`anonymousUser` and will thus cause a `UserNotFoundException` as this user does
not exist. This case has been handled in `ranking()` in `HomeController`.
This closes #98.