Update 'MatchControllerTest' to properly test 'MatchController'
Previously it was only tested whether the error messages would appear as the
users have never been sent to the Controller.
This has been fixed by accessing the nested `userId` field and setting it on
the neccessary fields.
The tests now also check if the controller methods properly handle the
exceptions thrown by `MatchService`