Change check for logged in user
Previously it was checked whether a user is logged in or not by
comparing the name to 'anonymousUser'. This was problematic while
testing and in the off chance a user with this name would register.
To circumvent this we now check whether a user is authenticated or not
by using `sec:authorize="isAuthenticated()"` which is part of the
`thymeleaf-extras-springsecurity` dependency.
This has replaced the old check in `header.html` and `ranking.html`
which should have been all uses unless any were overlooked.
With this change the tests had to be updated because due to the old
setup the tests would fail as security was disabled during the tests.
So to make the tests work security was enabled which caused the tests
to fail again because Keycloak would cause a `NullPointerException` due
to `AdapterConfig` null. To fix this all tests for the web layer now
import `KeycloakAutoConfiguration.class` to properly initialize
Keycloak in the Spring Context for the tests.