Removed all circular dependencies
Mostly it is just packages being moved around. The exception to this
was the "ranking" package. Due to it's close relation to the "user"
package the only easy fix to this was removing the "ranking" package and
making it part of the "user" package. This has solved the dependency
and should break nothing in the frontend as the fields remained the
same.