Provide backend for seasonal statistics
The old statistics are save in an extra table. Following that all
statistics are reset and all Matches are deleted. Any remaining
notifications should be deleted as well.
There is currently no option to query any of the seasons but that should
be easy to implement when necessary.
When the season has ended a "SeasonEndedEvent" is sent out. This should
allow us to for example send emails when a season has ended.